requirements.txt
flask==3.0.0
Pillow==10.1.0
tqdm==4.66.1
python-dotenv==1.0.0
werkzeug==3.0.1